Move mouseManager to a more often accessible file. Added more externs and function definitions
